💥 Fix excessive caching
Fix #971.

The old behavior for symbol declarations in v3 was that a public symbol declaration does not need to be imported for it to be included in the scope,
which means that a scoreboard objective `foo` referenced in the file `data/spgoding/functions/foo.mcfunction` could have a declaration at a completely unrelated file.
This requires the project to be checked twice to provide accurate error information,
as there is no way for the checkers to know if a symbol has a declaration elsewhere during the first pass.
To make the second pass faster,
the `TextDocument`s and ASTs of all files produced in the first pass are cached so that the second pass does not to redo all of the work
-- this, however, results in huge memory consumption (#971) as ASTs can apparently get quite big for large files (e.g. 60~ MB for a JSON file with 6~ million characters).

Instead of reading, parsing, and binding all files in the first pass, caching all the results, and checking them in the second pass,
this commit changes the behavior to reading, parsing, binding, and checking files in sequence in a single pass,
with the ASTs completely dropped from the memory after the files have been checked (except for `TextDocument`s managed by the client).

Due to this change, the file containing the declaration of a symbol now must be bound before a file using the symbol is bound for the symbol checker to be accurate.
This can be ensured in several ways:

* The vast majority of symbols in a data pack -- resource locations of Minecraft resources and other data pack files -- are bound by symbol registrars and URI binders in very early stage, so all references of them can be correctly bound.
* The `mcdoc` language already has proper import mechanisms and dispatchers don't need declarations, so they are not affected by this change.
* For symbols in `mcfunction` files that require declaration (e.g. bossbars, scoreboard objectives, teams, etc.), further improvement will be made to check `mcfunction` files starting from the `#minecraft:load` and `#minecraft:tick` function tags with every `function` command seen as an "import" directive to ensure most of those symbols are bound before they are referenced.
* For other symbols in `mcfunction` files that may be declared using Spyglass' custom `#declare` comment directives, users will need to put those comments strategically to make sure they are bound before symbol references. A new custom comment directive `#reference <category> <resource_location>` may also get added to inform the binder to bind the referenced resource first.