🐛 Allow single quotes in NBT paths
Fixes #1058. This change allows single quotes for
all versions, not just the versions after 1.20 Pre-
release 2, as the `nbt` package does not have
access to the `ReleaseVersion.cmp()` function from
`java-edition` and it'd take too much effort to
do it properly just to eliminate a tiny little
false negative. We may revisit this if we decide
to use pack formats internally instead since it'd
be an easy
`parseInt(ctx.project['loadedFormat']) < 15` check.