🐛 Use custom data instead of SymbolUtil alias
Alias requires the target symbol to exist and can
dereference multi-layer aliasing. We simply do not
need the first requirement as the target file may
not have been loaded yet by the time the local
binding is referenced, and we do not need the
second feature since use statement bindings are
local and cannot be used for imports across files.