🐛 Fix completion inside unclosed records
The code used to assume that the first and last characters in the range
of a record node are the brackets. No completions would be suggested for
key-value pairs when the cursor is at either end of the range, as that
is assumed to be outside the brackets. However, this assumption is
violated when there is no closing bracket, which is a common occurrence
during editing for users who do not rely on their editor to
automatically close brackets. This commit fixes the issue by keeping
track of the actual range inside the brackets.