🧱 Fix not generating SSH key for api-server user
The API server uses SSH for cloning as the mcmeta repo is too huge
for cloning/fetching with HTTPS. I think I generated the SSH keys once
manually and forgot to update the playbook.

The public SSH key will be printed to STDOUT after it's generated, and
needs to be added manually to some GitHub account; doesn't matter which
one since both vanilla-mcdoc and mcmeta are public. I've added it to
@SpyglassCrafter.