🩹 Fix sequence
This change allows the `gapParser` to be used to show context errors, like "expected space", between arguments, while also allowing for optional arguments in a sequence with such a `gapParser`.

The current behavior is gap parser will run between every item in the sequence, whether or not those parsers actually moved the cursor or returned nodes. So multiple chained optional arguments are fiddley to implement.